Fix a bug in anonymous functions’ type annotations
We cannot write anonymous functions with return type,
because it will cause parsing ambiguity. See
https://github.com/JuliaLang/julia/issues/23072.

+++ src/EoS.jl
@@ -53,7 +53,7 @@ end
 function eval_energy(eos::Birch)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         x = (v0 / v)^(2 / 3) - 1
         xi = 9 / 16 * b0 * v0 * x^2
         return f0 + 2 * xi + (bp0 - 4) * xi * x
@@ -63,7 +63,7 @@ end
 function eval_pressure(eos::Birch)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         x = v0 / v
         xi = x^(2 / 3) - 1
         return 3 / 8 * b0 * x^(5 / 3) * xi * (4 + 3 * (bp0 - 4) * xi)
@@ -73,7 +73,7 @@ end
 function eval_energy(eos::Murnaghan)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         x = bp0 - 1
         y = (v0 / v)^bp0
         return f0 + b0 / bp0 * v * (y / x + 1) - v0 * b0 / x
@@ -83,7 +83,7 @@ end
 function eval_pressure(eos::Murnaghan)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         return b0 / bp0 * ((v0 / v)^bp0 - 1)
     end
 end
@@ -91,7 +91,7 @@ end
 function eval_energy(eos::BirchMurnaghan2nd)::Function
     v0, b0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         f = ((v0 / v)^(2 / 3) - 1) / 2
         return f0 + 9 / 2 * b0 * v0 * f^2
     end
@@ -100,7 +100,7 @@ end
 function eval_pressure(eos::BirchMurnaghan2nd)::Function
     v0, b0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         f = ((v0 / v)^(2 / 3) - 1) / 2
         return 3 * b0 * f * (1 + 2 * f)^(5 / 2)
     end
@@ -109,7 +109,7 @@ end
 function eval_energy(eos::BirchMurnaghan3rd)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         eta = (v0 / v)^(1 / 3)
         xi = eta^2 - 1
         return f0 + 9 / 16 * b0 * v0 * xi^2 * (6 + bp0 * xi - 4 * eta^2)
@@ -119,7 +119,7 @@ end
 function eval_pressure(eos::BirchMurnaghan3rd)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         eta = (v0 / v)^(1 / 3)
         return 3 / 2 * b0 * (eta^7 - eta^5) * (1 + 3 / 4 * (bp0 - 4) * (eta^2 - 1))
     end
@@ -128,7 +128,7 @@ end
 function eval_energy(eos::BirchMurnaghan4th)::Function
     v0, b0, bp0, bpp0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         f = ((v0 / v)^(2 / 3) - 1) / 2
         h = b0 * bpp0 + bp0^2
         return f0 + 3 / 8 * v0 * b0 * f^2 * ((9 * h - 63 * bp0 + 143) * f^2 + 12 * (bp0 - 4) * f + 12)
@@ -138,7 +138,7 @@ end
 function eval_pressure(eos::BirchMurnaghan4th)::Function
     v0, b0, bp0, bpp0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         f = ((v0 / v)^(2 / 3) - 1) / 2
         h = b0 * bpp0 + bp0^2
         return 1 / 2 * b0 * (2 * f + 1)^(5 / 2) * ((9 * h - 63 * bp0 + 143) * f^2 + 9 * (bp0 - 4) * f + 6)
@@ -148,7 +148,7 @@ end
 function eval_energy(eos::Vinet)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64, f0::Float64=0)::Float64
+    function (v::Float64, f0::Float64=0)
         x = (v / v0)^(1 / 3)
         xi = 3 / 2 * (bp0 - 1)
         return f0 + 9 * b0 * v0 / xi^2 * (1 + (xi * (1 - x) - 1) * np.exp(xi * (1 - x)))
@@ -158,7 +158,7 @@ end
 function eval_pressure(eos::Vinet)::Function
     v0, b0, bp0 = eos.parameters
 
-    function (v::Float64)::Float64
+    function (v::Float64)
         x = (v / v0)^(1 / 3)
         xi = 3 / 2 * (bp0 - 1)
         return 3 * b0 / x^2 * (1 - x) * np.exp(xi * (1 - x))
